backend/requirements.txt
flask
tensorflow==2.15.0 
flask-cors
pillow
numpy